Move some tests around
