+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- The `http_server` input now provides a metadata field `http_server_request_path`.
+
 ## 3.43.1 - 2021-04-05
 
 ### Fixed
